pyproject.toml
[project]
name = "scaflog-zoho-mcp-server"
version = "0.1.0"
description = "MCP Server for Zoho Creator Integration"
readme = "README.md"
authors = [
    { name = "alexsherin", email = "alex.sherin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.1.0",
    "zcrmsdk==3.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.7.0",
    "httpx>=0.24.0",
]

[project.scripts]
scaflog-zoho-mcp-server = "scaflog_zoho_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scaflog_zoho_mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]